+++ src/native-lib/TuxedoECAPI.ts
@@ -1,0 +1,6 @@
+export interface ITuxedoECAPI {
+    webcamOn(): boolean;
+    webcamOff(): boolean;
+}
+
+export const TuxedoECAPI: ITuxedoECAPI = require('TuxedoECAPI');