+++ src/ng-app/app/config.service.ts
@@ -6,7 +6,7 @@ import { ITccProfile } from '../../common/models/TccProfile';
 import { ConfigHandler } from '../../common/classes/ConfigHandler';
 import { environment } from '../environments/environment';
 import { ElectronService } from 'ngx-electron';
-import { Observable, Subject } from 'rxjs';
+import { Observable, Subject, BehaviorSubject } from 'rxjs';
 import { UtilsService } from './utils.service';
 
 @Injectable({
@@ -28,28 +28,33 @@ export class ConfigService {
 
   public observeEditingProfile: Observable<ITccProfile>;
   private editingProfileSubject: Subject<ITccProfile>;
+  public editingProfile: BehaviorSubject<ITccProfile>;
 
   // Exporting of relevant functions from ConfigHandler
   // public copyConfig = ConfigHandler.prototype.copyConfig;
   // public writeSettings = ConfigHandler.prototype.writeSettings;
 
-  constructor(private electron: ElectronService, utils: UtilsService) {
+  constructor(private electron: ElectronService, private utils: UtilsService) {
     this.settingsSubject = new Subject<ITccSettings>();
     this.observeSettings = this.settingsSubject.asObservable();
 
     this.editingProfileSubject = new Subject<ITccProfile>();
     this.observeEditingProfile = this.editingProfileSubject.asObservable();
+    this.editingProfile = new BehaviorSubject<ITccProfile>(undefined);
 
     this.config = new ConfigHandler(TccPaths.SETTINGS_FILE, TccPaths.PROFILES_FILE, TccPaths.AUTOSAVE_FILE);
     this.defaultProfiles = this.config.getDefaultProfiles();
     for (const profile of this.defaultProfiles) {
-      utils.fillDefaultValuesProfile(profile);
+      this.utils.fillDefaultValuesProfile(profile);
     }
     this.readFiles();
   }
 
   public readFiles(): void {
     this.customProfiles = this.config.getCustomProfilesNoThrow();
+    for (const profile of this.customProfiles) {
+      this.utils.fillDefaultValuesProfile(profile);
+    }
     this.settings = this.config.getSettingsNoThrow();
     this.settingsSubject.next(this.settings);
   }
@@ -140,6 +145,79 @@ export class ConfigService {
     }
   }
 
+  private async pkexecWriteCustomProfilesAsync(customProfiles: ITccProfile[]): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      const tmpProfilesPath = '/tmp/tmptccprofiles';
+      this.config.writeProfiles(customProfiles, tmpProfilesPath);
+      let tccdExec: string;
+      if (environment.production) {
+        tccdExec = TccPaths.TCCD_EXEC_FILE;
+      } else {
+        tccdExec = this.electron.process.cwd() + '/dist/tuxedo-control-center/data/service/tccd';
+      }
+      this.utils.execCmd('pkexec ' + tccdExec + ' --new_profiles ' + tmpProfilesPath).then(data => {
+        resolve(true);
+      }).catch(error => {
+        resolve(false);
+      });
+    });
+  }
+
+  public async writeProfile(currentProfileName: string, profile: ITccProfile, states?: string[]): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      const profileIndex = this.customProfiles.findIndex(p => p.name === currentProfileName);
+
+      // Copy custom profiles and if provided profile is one of them, overwrite with
+      // provided profile
+      const customProfilesCopy = this.config.copyConfig<ITccProfile[]>(this.customProfiles);
+      if (profileIndex !== -1) {
+        customProfilesCopy[profileIndex] = profile;
+      } else {
+        if (this.defaultProfiles.find(p => p.name === currentProfileName) === undefined) {
+          resolve(false);
+          return;
+        }
+      }
+
+      // Copy config and if states are provided, assign the chosen profile to these states
+      const newSettings: ITccSettings = this.config.copyConfig<ITccSettings>(this.getSettings());
+      if (states !== undefined) {
+        for (const stateId of states) {
+          newSettings.stateMap[stateId] = profile.name;
+        }
+      }
+
+      this.pkexecWriteConfigAsync(newSettings, customProfilesCopy).then(success => {
+        if (success) {
+          this.readFiles();
+        }
+        resolve(success);
+      });
+    });
+  }
+
+  private async pkexecWriteConfigAsync(settings: ITccSettings, customProfiles: ITccProfile[]): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      const tmpProfilesPath = '/tmp/tmptccprofiles';
+      const tmpSettingsPath = '/tmp/tmptccsettings';
+      this.config.writeProfiles(customProfiles, tmpProfilesPath);
+      this.config.writeSettings(settings, tmpSettingsPath);
+      let tccdExec: string;
+      if (environment.production) {
+        tccdExec = TccPaths.TCCD_EXEC_FILE;
+      } else {
+        tccdExec = this.electron.process.cwd() + '/dist/tuxedo-control-center/data/service/tccd';
+      }
+      this.utils.execCmd(
+        'pkexec ' + tccdExec + ' --new_profiles ' + tmpProfilesPath + ' --new_settings ' + tmpSettingsPath
+      ).then(data => {
+        resolve(true);
+      }).catch(error => {
+        resolve(false);
+      });
+    });
+  }
+
   /**
    * Retrieves the currently chosen profile for edit
    *
@@ -192,6 +270,7 @@ export class ConfigService {
       this.currentProfileEditIndex = -1;
       this.currentProfileEdit = undefined;
       this.editingProfileSubject.next(undefined);
+      this.editingProfile.next(undefined);
     }
     const index = this.currentProfileEditIndex = this.customProfiles.findIndex(e => e.name === customProfileName);
     if (index === -1) {
@@ -199,7 +278,9 @@ export class ConfigService {
     } else {
       this.currentProfileEditIndex = index;
       this.currentProfileEdit = this.config.copyConfig<ITccProfile>(this.customProfiles[index]);
+      this.utils.fillDefaultValuesProfile(this.currentProfileEdit);
       this.editingProfileSubject.next(this.currentProfileEdit);
+      this.editingProfile.next(this.currentProfileEdit);
       return true;
     }
   }